Move the content of 'script/style' into the new 'style' crate.
