+++ src/components/style/properties/.gitignore
