+++ src/components/style/style.rc
@@ -11,4 +11,25 @@
 #[license = "MPL"];
 #[crate_type = "lib"];
 
+
+extern mod extra;
 extern mod cssparser;
+extern mod script;
+
+
+// The "real" public API
+pub use self::selector_matching::{Stylist, StylesheetOrigin};
+
+
+// Things that need to be public to make the compiler happy
+pub mod stylesheets;
+pub mod errors;
+pub mod selectors;
+pub mod selector_matching;
+pub mod properties;
+pub mod namespaces;
+pub mod media_queries;
+pub mod parsing_utils;
+
+#[cfg(test)]
+mod tests;
